Major commit to reorgansise code into a modular menu
now the main pp6calculator just calls the top menu which in turn calls
the weekOneTopMenu which contains the code to do the first week
operations this will be further improved to give each operation a
function which is called from the menu. making the menu if list easier
to read.

onece the weekOneMenu is exited it should return to the topMenu which
will let you reselect the menu you whish to use or exit the program
entierly